begin separation of constraints and modifiers
