+++ pyhf/modifiers/histosys.py
@@ -4,6 +4,7 @@ log = logging.getLogger(__name__)
 from . import modifier
 from .. import get_backend
 from ..interpolate import interpolator
+from ..constraints import standard_constraint
 
 @modifier(name='histosys', constrained=True, shared=True, op_code = 'addition')
 class histosys(object):
@@ -24,18 +25,13 @@ class histosys(object):
         self.at_minus_one = {}
         self.at_plus_one = {}
         self.auxdata = [0]  # observed data is always at a = 1
+        self.constraint = standard_constraint()
 
     def add_sample(self, channel, sample, modifier_def):
         log.info('Adding sample {0:s} to channel {1:s}'.format(sample['name'], channel['name']))
         self.at_zero.setdefault(channel['name'], {})[sample['name']] = sample['data']
         self.at_minus_one.setdefault(channel['name'], {})[sample['name']] = modifier_def['data']['lo_data']
         self.at_plus_one.setdefault(channel['name'], {})[sample['name']] = modifier_def['data']['hi_data']
-
-    def alphas(self, pars):
-        return pars  # the nuisance parameters correspond directly to the alpha
-
-    def expected_data(self, pars):
-        return self.alphas(pars)
 
     def apply(self, channel, sample, pars):
         assert int(pars.shape[0]) == 1
