+++ pyhf/modifiers/normsys.py
@@ -4,6 +4,7 @@ log = logging.getLogger(__name__)
 from . import modifier
 from .. import get_backend
 from ..interpolate import interpolator
+from ..constraints import standard_constraint
 
 @modifier(name='normsys', constrained=True, shared=True, op_code = 'multiplication')
 class normsys(object):
@@ -17,17 +18,13 @@ class normsys(object):
         self.at_minus_one = {}
         self.at_plus_one = {}
         self.auxdata = [0]  # observed data is always at a = 1
+        self.constraint = standard_constraint()
 
     def add_sample(self, channel, sample, modifier_def):
         log.info('Adding sample {0:s} to channel {1:s}'.format(sample['name'], channel['name']))
         self.at_minus_one.setdefault(channel['name'], {})[sample['name']] = [modifier_def['data']['lo']]
         self.at_plus_one.setdefault(channel['name'], {})[sample['name']]  = [modifier_def['data']['hi']]
 
-    def alphas(self, pars):
-        return pars  # the nuisance parameters correspond directly to the alpha
-
-    def expected_data(self, pars):
-        return self.alphas(pars)
 
     def apply(self, channel, sample, pars):
         # normsysfactor(nom_sys_alphas)   = 1 + sum(interp(1, anchors[i][0], anchors[i][0], val=alpha)  for i in range(nom_sys_alphas))
