+++ pyhf/pdf.py
@@ -356,7 +356,7 @@ class Model(object):
         # order matters! because we generated auxdata in a certain order
         auxdata = None
         for modname in self.config.auxdata_order:
-            thisaux = self.config.modifier(modname).expected_data(
+            thisaux = self.config.modifier(modname).constraint.expected_data(
                 pars[self.config.par_slice(modname)])
             tocat = [thisaux] if auxdata is None else [auxdata, thisaux]
             auxdata = tensorlib.concatenate(tocat)
